Create API endpoint for OLC paths. Closes #4 .
